wrote functions for calculating dimensions, cell volumes and cell centers
